WIP #85: Further cleanups and get rid of pathlib due to poor symlink support :-(

+++ workflows/utils.py
@@ -2,14 +2,14 @@
 #
 # Author: Yipeng Sun
 # License: BSD 2-clause
-# Last Change: Mon Oct 04, 2021 at 05:23 PM +0200
+# Last Change: Mon Oct 04, 2021 at 10:26 PM +0200
 
 import re
 import yaml
-import os
-import os.path as os_path
+import shlex
+import os.path as op  # Note: Can't use pathlib because that doesn't handle symbolic link well
 
-from os import makedirs, chdir, symlink, getcwd, environ, pathsep
+from os import makedirs, chdir, symlink, getcwd, system
 from datetime import datetime
 from subprocess import check_output
 from shutil import rmtree
@@ -18,74 +18,40 @@ from glob import glob
 from pyBabyMaker.base import TermColor as TC
 
 
-###################
-# Generic helpers #
-###################
+################
+# Path helpers #
+################
 
-def gen_date(fmt='%y_%m_%d'):
-    return datetime.now().strftime(fmt)
-
-
-def load_yaml_db(yaml_db=[
-        'rdx/rdx-run2.yml',
-        'rdx/rdx-run1.yml',
-]):
-    result = dict()
-    base_path = os_path.abspath(os_path.dirname(__file__))
-
-    for db in yaml_db:
-        with open(os_path.join(base_path, db)) as f:
-            result.update(yaml.safe_load(f))
-
-    return result
+def abs_path(path, base_path=op.abspath(op.dirname(__file__))):
+    return op.abspath(op.join(base_path, path))
 
 
-def run_cmd_wrapper(only_print=False):
-    def inner(cmd):
-        print('  {}'.format(cmd))
-
-        if not only_print:
-            os.system(cmd)
-    return inner
-
-
-###############
-# I/O helpers #
-###############
-
-def abs_path(path, base_path=__file__):
-    return os_path.abspath(
-        os_path.join(os_path.dirname(os_path.abspath(base_path)), path))
+def with_suffix(path, ext):
+    return op.splitext(path)[0] + ext
 
 
 def ensure_dir(path, delete_if_exist=True, **kwargs):
     path = abs_path(path, **kwargs)
 
-    if os_path.isdir(path):
-        if delete_if_exist:
-            rmtree(path)
-            makedirs(path)
+    if delete_if_exist and op.isdir(path):
+        rmtree(path)
 
-    else:
-        makedirs(path)
-
+    makedirs(path)
     return path
 
 
-def find_all_input(inputs, patterns=['*.root']):
+def find_all_input(inputs, patterns=['*.root'], blocked_patterns=['__aux']):
     result = []
     for f in inputs:
-        if os_path.isfile(f):
-            result.append(os_path.abspath(f))
-        elif os_path.isdir(f):
+        if op.isfile(f):
+            result.append(op.abspath(f))
+        elif op.isdir(f):
             for p in patterns:
-                result += [os_path.abspath(g) for g in glob(os_path.join(f, p))]
-    return result
+                result += [op.abspath(g) for g in glob(op.join(f, p))]
 
+    # Remove files that contains blocked patterns
+    return [f for f in result if True not in [p in f for p in blocked_patterns]]
 
-#####################
-# Execution helpers #
-#####################
 
 def aggragate_output(workdir, output_dir, keep):
     # Symbolic link generated files that match 'keep' patterns in separate
@@ -93,19 +59,58 @@ def aggragate_output(workdir, output_dir, keep):
 
     # NOTE: 'workdir' is usually the main workdir of the fulljob,
     #       'output_dir' the workdir of a subjob
-    workdir = os_path.abspath(workdir)
-    chdir(os_path.abspath(workdir))
-    output_dir = os_path.abspath(output_dir)
+    workdir = op.abspath(workdir)
+    chdir(op.abspath(workdir))
+    output_dir = op.abspath(output_dir)
 
     for d, patterns in keep.items():
         chdir(workdir)
         ensure_dir(d, False)
         chdir(d)
-        relpath = os_path.relpath(output_dir, os_path.abspath('.'))
+        relpath = op.relpath(output_dir, op.abspath('.'))
 
         for p in patterns:
-            for f in glob(os_path.join(relpath, p)):
-                symlink(f, os_path.join('.', os_path.basename(f)))
+            for f in glob(op.join(relpath, p)):
+                symlink(f, op.join('.', op.basename(f)))
+
+
+################
+# MISC helpers #
+################
+
+def gen_date(fmt='%y_%m_%d'):
+    return datetime.now().strftime(fmt)
+
+
+def load_yaml_db(yaml_db=[
+        'rdx/rdx-run1.yml',
+        'rdx/rdx-run2.yml',
+], base_path=abs_path('.')):
+    result = dict()
+
+    for db in yaml_db:
+        with open(op.join(base_path, db)) as f:
+            result.update(yaml.safe_load(f))
+
+    return result
+
+
+#####################
+# Execution helpers #
+#####################
+
+def run_cmd_wrapper(only_print=False):
+    def inner(cmd):
+        print('  {}'.format(cmd))
+
+        if not only_print:
+            system(cmd)
+    return inner
+
+
+def run_cmd_with_output(cmd):
+    cmd_splitted = shlex.split(cmd)
+    return check_output(cmd_splitted).decode('utf-8').strip()
 
 
 ################################
@@ -157,3 +162,10 @@ def parse_step2_name(ntp_name):
 #####################
 # Generic workflows #
 #####################
+
+def wf_compile_cpp(input_cpp, add_flags='-I{}'.format(abs_path('../include'))):
+    compiler = run_cmd_with_output('root-config --cxx')
+    base_flags = run_cmd_with_output('root-config --cflags')
+    link_flags = run_cmd_with_output('root-config --libs')
+
+    # output_exe = Path(input_cpp).bas
